% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.all_have_DNAStringSet}
\alias{.all_have_DNAStringSet}
\title{test all TSEs have DNAStringSet in the referenceSeq slot}
\usage{
.all_have_DNAStringSet(args)
}
\description{
test all TSEs have DNAStringSet in the referenceSeq slot
}
\keyword{internal}
