% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.all_have_DNAStringSetList}
\alias{.all_have_DNAStringSetList}
\title{test all TSEs have DNAStringSetList in the referenceSeq slot}
\usage{
.all_have_DNAStringSetList(args)
}
\description{
test all TSEs have DNAStringSetList in the referenceSeq slot
}
\keyword{internal}
