% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.auto_rename_list}
\alias{.auto_rename_list}
\title{rename a list automatically to avoid duplicated names}
\usage{
.auto_rename_list(x)
}
\description{
rename a list automatically to avoid duplicated names
}
\keyword{internal}
