% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.bind_link_tree}
\alias{.bind_link_tree}
\title{bind links & trees when combine TSE}
\usage{
.bind_link_tree(x, args, drop.rowLinks, drop.colLinks, bind = "cbind")
}
\description{
bind links & trees when combine TSE
}
\keyword{internal}
