% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classValid.R
\name{.checkLDF}
\alias{.checkLDF}
\title{\code{validLDF} is to valid a \code{LinkDataFrame} object.}
\usage{
.checkLDF(object)
}
\arguments{
\item{object}{A \code{LinkDataFrame} object}
}
\value{
TRUE or a character string.
}
\description{
\code{validLDF} is to valid a \code{LinkDataFrame} object.
}
\keyword{internal}
