% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classValid.R
\name{.checkTSE}
\alias{.checkTSE}
\title{valid TreeSummarizedExperiment class}
\usage{
.checkTSE(object)
}
\arguments{
\item{object}{A TreeSummarizedExperiment object}
}
\value{
TRUE or a character string.
}
\description{
\code{validTSE} is to valid TreeSummarizedExperiment object.
}
\keyword{internal}
