% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.combine_link_tree}
\alias{.combine_link_tree}
\title{update links & trees when combine TSE}
\usage{
.combine_link_tree(x, args, drop.rowLinks, drop.colLinks, bind = "cbind")
}
\description{
update links & trees when combine TSE
}
\keyword{internal}
