% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.is_equal_link}
\alias{.is_equal_link}
\title{The links & trees in the specified dim are consistent}
\usage{
.is_equal_link(args, dim = "row")
}
\description{
The links & trees in the specified dim are consistent
}
\keyword{internal}
