% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.match_phylo_list}
\alias{.match_phylo_list}
\title{match a list of phylo (x.phys) against to a list of phylo (y.phys)}
\usage{
.match_phylo_list(x.phys, y.phys)
}
\description{
match a list of phylo (x.phys) against to a list of phylo (y.phys)
}
\keyword{internal}
