% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.match_x_dupY}
\alias{.match_x_dupY}
\title{convert char. indicator to num. indicator}
\usage{
.match_x_dupY(x, dy)
}
\arguments{
\item{x}{A vector. The values to be matched.}

\item{dy}{A vector. The values to be matched agaist.}
}
\description{
This differs to \code{match} with that the duplicated values in dy are not
ignored.
}
\author{
Ruizhu Huang
}
\keyword{internal}
