% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.name_y_with_x}
\alias{.name_y_with_x}
\title{name y with x}
\usage{
.name_y_with_x(x, y)
}
\description{
name y with x
}
\examples{
\dontrun{
x <- letters[1:5]
y <- 1:5
.name_y_with_x(x, y)
}
}
\keyword{internal}
