% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.numeric_ij}
\alias{.numeric_ij}
\title{convert char. indicator to num. indicator}
\usage{
.numeric_ij(ij, x, dim = "row")
}
\arguments{
\item{ij}{A character or numeric indicator on rows/columns of \code{x}}

\item{x}{It provides row/col names for \code{ij} to be matched against.}

\item{dim}{"row" or "col" to specify row/col names of \code{x} to be matched
against.}
}
\description{
convert char. indicator to num. indicator
}
\author{
Ruizhu Huang
}
\keyword{internal}
