% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.replace_link_tree_1d}
\alias{.replace_link_tree_1d}
\title{replace row/col links & trees}
\usage{
.replace_link_tree_1d(x, value, ij, dim = "row")
}
\arguments{
\item{x}{A TSE with \code{ij} rows/cols to be replaced by \code{value}}

\item{value}{A TSE to replace some rows/cols of \code{x}.}

\item{ij}{A character or numeric vector to specify which rows/cols to be replaced.}

\item{dim}{"row" or "col" to specify the dimension is in rows or columns}
}
\description{
replace row/col links & trees
}
\author{
Ruizhu Huang
}
\keyword{internal}
