% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.subset_leaf}
\alias{.subset_leaf}
\title{update dimLinks and dimTree (used in subsetByLeaf)}
\usage{
.subset_leaf(x, leaf, dim = "row", updateTree = TRUE)
}
\description{
update dimLinks and dimTree (used in subsetByLeaf)
}
\author{
Ruizhu Huang
}
\keyword{internal}
