% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{.update_whichTree}
\alias{.update_whichTree}
\title{update the 'whichTree' column in row/column link data}
\usage{
.update_whichTree(x, y)
}
\description{
update the 'whichTree' column in row/column link data
}
\examples{
\dontrun{
(ld <- LinkDataFrame(nodeLab = letters[1:5],
                     nodeLab_alias = LETTERS[1:5],
                     nodeNum = 1:5,
                     isLeaf = TRUE,
                     whichTree = LETTERS[1:5],
                     right = 1:5))
newWhich <- setNames(letters[1:5], LETTERS[1:5])
.update_whichTree(ld, y = newWhich)
}
}
\keyword{internal}
