% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/classAccessor.R
\docType{methods}
\name{[,LinkDataFrame,ANY,ANY,ANY-method}
\alias{[,LinkDataFrame,ANY,ANY,ANY-method}
\alias{as.data.frame,LinkDataFrame-method}
\alias{$,LinkDataFrame-method}
\alias{$<-,LinkDataFrame-method}
\alias{LinkDataFrame-accessor}
\title{LinkDataFrame-accessor}
\usage{
\S4method{[}{LinkDataFrame,ANY,ANY,ANY}(x, i, j)

\S4method{as.data.frame}{LinkDataFrame}(x)

\S4method{$}{LinkDataFrame}(x, name)

\S4method{$}{LinkDataFrame}(x, name) <- value
}
\arguments{
\item{x}{A \code{LinkDataFrame} object}

\item{i, j}{The row, column index to subset \code{x}.}

\item{name}{The name of the column.}

\item{value}{The value to replace.}
}
\value{
Depends on the functions
}
\description{
This page lists the accesssor, coercion, subsetting and combining functions
for the class \strong{LinkDataFrame}.
}
\examples{

left <- DataFrame(left1 = 1:5, left2 = letters[1:5])
right <- DataFrame(right1 = sample(letters[1:3], 5, replace = TRUE),
                  right2 = sample(c(TRUE, FALSE), 5, replace = TRUE),
                  right3 = 11:15)

ld <- LinkDataFrame(LinkData = left, right)

## subset
# by rows
(ld1 <- ld[1:3, ])
# by columns (only on the right side of the vertical line)
(ld2 <- ld[, c(1, 3)])
(ld3 <- ld[, c("right1", "right2")])

## add a new column (only on the right side of the vertical line)
ld4 <- ld
ld4$right4 <- 11:15
ld4

## coercion
(ld5 <- as.data.frame(ld))


}
\seealso{
\code{\link{LinkDataFrame-constructor}}
  \code{\link{LinkDataFrame-class}} \code{\link[S4Vectors]{DataFrame}}
}
\author{
Ruizhu HUANG
}
