% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R
\docType{class}
\name{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment-class}
\alias{MicrobiomeExperiment}
\title{The \code{MicrobiomeExperiment} class}
\usage{
MicrobiomeExperiment(..., referenceSeq = NULL)
}
\arguments{
\item{...}{Arguments passed to
\code{\link[=TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}

\item{referenceSeq}{a \code{DNAStringSet} object or some object
coercible to a \code{DNAStringSet} object. See
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}} for more
details.}
}
\description{
The \code{MicrobiomeExperiment} class is a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}-like
class for microbiome data inheriting from
\code{\link[=TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
and
\code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}}.
}
\details{
\code{MicrobiomeExperiment} adds to the above mentioned class a slot for
storing a reference sequence as a
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}}.
}
\examples{
sampleNames <- letters[1:4]
cd <- DataFrame(a = letters[1:4], b = 1:4)
rd <- DataFrame(Kingdom = "Bacteria",
                Phylum = c("Firmicutes","Firmicutes","Firmicutes",
                           "Bacteroidetes","Euryarchaeota"),
                Class = c("Bacilli","Bacilli","Bacilli","Saprospirae",
                          "Methanobacteria"),
                Order = c("Bacillales","Lactobacillales","Lactobacillales",
                          "Saprospirales","Methanobacteriales"),
                Family = c("Planococcaceae","Enterococcaceae","Enterococcaceae",
                           "Chitinophagaceae","Methanobacteriaceae"),
                Genus = c("Staphylococcus",NA,"Melissococcus",
                          "Sediminibacterium","Methanobrevibacter"))
counts <- matrix(sample(1:1000, nrow(rd) * nrow(cd), replace=TRUE),
                 nr = nrow(rd),
                 nc = nrow(cd))
refSeq <- DNAStringSetList(one = DNAStringSet(c("A","A","A","A","A")),
                           two = DNAStringSet(c("A","A","A","A","A")))
me <- MicrobiomeExperiment(assays = SimpleList(counts = counts),
                           rowData = rd,
                           colData = cd,
                           referenceSeq = refSeq)
me
}
