% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R,
%   R/TreeSummarizedExperiment.R
\name{vertical_slot_names,MicrobiomeExperiment-method}
\alias{vertical_slot_names,MicrobiomeExperiment-method}
\alias{[,MicrobiomeExperiment,ANY,ANY,ANY-method}
\alias{[<-,MicrobiomeExperiment,ANY,ANY,MicrobiomeExperiment-method}
\alias{MicrobiomeExperiment-internal}
\title{MicrobiomeExperiment internals}
\usage{
\S4method{vertical_slot_names}{MicrobiomeExperiment}(x)

\S4method{[}{MicrobiomeExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{MicrobiomeExperiment,ANY,ANY,MicrobiomeExperiment}(x, i, j, ...) <- value
}
\description{
Internal functions which should only be used in MicrobiomeExperiment
}
\keyword{internal}
