% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeValue.R
\docType{methods}
\name{nodeValue}
\alias{nodeValue}
\alias{nodeValue,ANY-method}
\alias{nodeValue,treeSummarizedExperiment-method}
\title{Calculate entity values at internal nodes}
\usage{
nodeValue(data, fun = sum, tree, level = NULL, message = FALSE)

\S4method{nodeValue}{ANY}(data, fun = sum, tree, level = NULL,
  message = FALSE)

\S4method{nodeValue}{treeSummarizedExperiment}(data, fun = sum,
  level = NULL, message = FALSE)
}
\arguments{
\item{data}{A leafSummarizedExperiment object or a data frame/matrix.}

\item{fun}{A function to create the value of an internal node based on values
at its descendant leaf nodes. The default is sum.}

\item{tree}{An optional argument. Only need if \code{data} is a data frame or
matrix.}

\item{level}{Default is NULL. It could be a vector of nodes using node value
or label to specify the level the data should be aggregated to.}

\item{message}{A logical value. The default is TRUE. If TRUE, it will print
out the currenet status of a process.}
}
\value{
If input data is a treeSummarizedExperiment object, a
  treeSummarizedExperiment is returned; otherwise, a matrix is returned
}
\description{
\code{nodeValue} calculates value, e.g., count, for each internal
node based on the values of its descendant leaves. For example, if the
abundance of a cell cluster, which corresponds to an internal node in the
tree, is of interest, we could sum the abundance of all cell types in that
cluster. Each cell type, in this case, corresponds to a leaf node in the
tree. Different calculations, instead of only sum, might be required. This
could be achieved by specifying different functions in the argument
\strong{fun}.
}
\examples{
data("tinyTree")
if (require(ggtree)) {
(p <- ggtree(tinyTree) + geom_text(aes(label = label)))
set.seed(1)
count <- matrix(rpois(100, 10), nrow =10)
rownames(count) <- tinyTree$tip.label
colnames(count) <- paste(c("C1_", "C2_"),
c(1:5, 1:5), sep = "")

count_tinyTree <- nodeValue(data = count,
tree = tinyTree, fun = sum)

# check to see whether the count of an internal node is the sum
# of counts of its descendant leaves.
# here, check the first sample as an example

nod <- transNode(tree = tinyTree, input = rownames(count_tinyTree),
                 message = FALSE)
d <- cbind.data.frame(node = nod, count = count_tinyTree[, 1])

ggtree(tinyTree) \%<+\% d + geom_text2(aes(label = count))


# aggregate to a specific level
count <- nodeValue(data = count, tree = tinyTree,
                   fun = sum, level = c(18, 12, 9))
}

}
\seealso{
\code{\link{treeSummarizedExperiment}} for treeSummarizedExperiment
creation.
}
\author{
Ruizhu Huang
}
