% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{data}
\name{phylo}
\alias{phylo}
\title{phylo: A S3 class for the phylogenetic tree}
\format{
An object of class \code{phylo} of length 0.
}
\usage{
phylo
}
\description{
The \pkg{ape} package does not export its phylo class, probably because it
is not really defined formally anywhere. Technically, it is an S3 class
extended from the class list. Any exported definitions from the \pkg{ape}
package would be preferred to use if available.
}
\keyword{internal}
