% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExperiment-class.R
\name{referenceSeq}
\alias{referenceSeq}
\alias{referenceSeq,MicrobiomeExperiment-method}
\alias{referenceSeq<-}
\alias{referenceSeq<-,MicrobiomeExperiment-method}
\title{Microbiome data methods}
\usage{
referenceSeq(x)

\S4method{referenceSeq}{MicrobiomeExperiment}(x)

referenceSeq(x) <- value

\S4method{referenceSeq}{MicrobiomeExperiment}(x) <- value
}
\arguments{
\item{x}{a \code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}}
object}

\item{value}{a a \code{\link[Biostrings:XStringSet-class]{DNAStringSet}}
object or an object coercible to one.}
}
\description{
Methods to get or set reference sequence data on a
\code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}} object.
}
