% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareNode.R
\name{shareNode}
\alias{shareNode}
\title{Find the share node}
\usage{
shareNode(tree, node, use.alias = FALSE)
}
\arguments{
\item{tree}{A  phylo object.}

\item{node}{A vector of node numbers or node labels.}

\item{use.alias}{A logical value, TRUE or FALSE. The default is FALSE, and
the node label would be used to name the output; otherwise, the alias of
node label would be used to name the output. The alias of node label is
created by adding a prefix \code{"Node_"} to the node number if the node is
an internal node or adding a prefix \code{"Leaf_"} if the node is a leaf
node.}
}
\value{
A vector of nodes. The numeric value is the node number, and the
  vector name is the corresponding node label. If a node has no label, it
  would have NA as name when \code{use.alias = FALSE}, and have the alias of
  node label as name when \code{use.alias = TRUE}.
}
\description{
\code{shareNode} is to find the node where the specified nodes first meet.
}
\examples{
library(ggtree)
data(tinyTree)

# PLOT tree
ggtree(tinyTree, branch.length = 'none') +
    geom_text2(aes(label = label), color = "darkorange",
           hjust = -0.1, vjust = -0.7) +
    geom_text2(aes(label = node), color = "darkblue",
               hjust = -0.5, vjust = 0.7)


## find the node shared by provided node labels
shareNode(node = c('t4','t9'), tree = tinyTree,
          use.alias = FALSE)

shareNode(node = c('t10','Node_17'), tree = tinyTree,
          use.alias = FALSE)

## find the node shared by provided node numbers
shareNode(node = c(2, 3), tree = tinyTree)
}
\author{
Ruizhu Huang
}
