% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\docType{methods}
\name{show,LinkDataFrame-method}
\alias{show,LinkDataFrame-method}
\title{The show method of the class \strong{LinkDataFrame} is modified based on the
codes of the show method of the class \strong{"GPos"}. The original code
could be found in the
https://github.com/Bioconductor/GenomicRanges/blob/master/R/GPos-class.R}
\usage{
\S4method{show}{LinkDataFrame}(object)
}
\description{
The show method of the class \strong{LinkDataFrame} is modified based on the
codes of the show method of the class \strong{"GPos"}. The original code
could be found in the
https://github.com/Bioconductor/GenomicRanges/blob/master/R/GPos-class.R
}
\keyword{internal}
