% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_toTree.R
\name{toTree}
\alias{toTree}
\title{Translate a data frame to a phylo object}
\usage{
toTree(data, edges = FALSE)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{edges}{A logical value, TRUE or FALSE. The default is FALSE. If TRUE,
the input data should be a matrix of edges (A matrix with two columns). The
first column is the parent and the second is the child. If FALSE,
\code{data} should be a taxonomic table (see examples).}
}
\value{
a phylo object
}
\description{
\code{toTree} translates a data frame to a phylo object
}
\details{
The last column is used as the leaf nodes
}
\examples{
# data is a taxonomic table
taxTab <- data.frame(R1 = rep("A", 5),
R2 = c("B1", rep("B2", 4)),
R3 = c("C1", "C2", "C3", "C3", "C4"))

taxTab <- data.frame(R1 = rep("A", 5),
R2 = c("B1", rep("B2", 2), NA, "B2"),
R3 = c("C1", "C2", "C3", NA, "C4"))

tree <- toTree(data = taxTab, edges = FALSE)


# data is a matrix of edges
# the matrix should be in character
 mat <- tree$edge
 mat <- apply(mat, 2, function(x){
          transNode(tree, node = x)})
 tree2 <- toTree(mat, edges = TRUE)
 
}
\author{
Ruizhu HUANG
}
