% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_transNode.R
\name{transNode}
\alias{transNode}
\title{Transfer between node number and node label}
\usage{
transNode(tree, input, use.alias = FALSE, message = FALSE)
}
\arguments{
\item{tree}{A phylo object}

\item{input}{A character or numeric vector representing tree node label(s) or
tree node number(s)}

\item{use.alias}{A logical value, TRUE or FALSE. This is an optional argument
that only requried when \code{return = "label"}. The default is FALSE, and
the node label would be returned; otherwise, the alias of node label would
be output. The alias of node label is created by adding a prefix
\code{"Node_"} to the node number if the node is an internal node or adding
a prefix \code{"Leaf_"} if the node is a leaf node.}

\item{message}{A logical value, TRUE or FALSE. The default is FALSE. If TRUE,
message will show when a tree have duplicated labels for some internal
nodes.}
}
\value{
a vector
}
\description{
\code{transNode} does the transformation between the number and the label of
a node on a tree
}
\examples{
library(ggtree)

data(tinyTree)

ggtree(tinyTree, branch.length = 'none') +
geom_text2(aes(label = label), hjust = -0.3) +
geom_text2(aes(label = node), vjust = -0.8,
hjust = -0.3, color = 'blue')

#check whether the node number and node label are matched
transNode(tinyTree, input = c(11, 2, 4, 15))

transNode(tree = tinyTree, input = c("Node_16", "Node_11"))

}
\author{
Ruizhu Huang
}
