% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treeC}
\alias{treeC}
\title{A simulated phylogenetic tree with 4 tips and 3 internal nodes}
\format{A phylo object with 4 tips and 3 internal nodes:
\describe{
  \item{Tip labels}{t1, t2, ..., t4.}
}}
\usage{
treeC
}
\description{
A random phylo object created using the function \code{\link[ape]{rtree}}
}
\keyword{datasets}
