% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperiment.R
\docType{package}
\name{TreeeSummarizedExperiment-package}
\alias{TreeeSummarizedExperiment-package}
\title{The \code{TreeSummarizedExperiment} package}
\description{
\code{TreeSummarizedExperiment} implement a class of the same name, which
extends SingleCellExperiment to include hierarchical information on the rows
or columns of the rectangular data.
}
\details{
It also includes an additional slot for storing reference sequences per
feature.
}
\seealso{
\link[=TreeSummarizedExperiment-class]{TreeSummarizedExperiment} class
}
