% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treeR}
\alias{treeR}
\title{A simulated phylogenetic tree with 5 tips and 4 internal nodes}
\format{A phylo object with 5 tips and 4 internal nodes:
\describe{
  \item{Tip labels}{t1, t2, ..., t5.}
}}
\usage{
treeR
}
\description{
A random phylo object created using the function \code{\link[ape]{rtree}}
}
\keyword{datasets}
