% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{TreeSummarizedExperiment-class}
\alias{TreeSummarizedExperiment-class}
\title{An S4 class TreeSummarizedExperiment}
\description{
The class \strong{TreeSummarizedExperiment} is an extension class of standard
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} class. It has
five slots. Four of them are traditional slots from
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} class:
\code{assays}, \code{rowData} \code{colData} and \code{metadata}. The new
slot is \code{treeData} that is to store the hiearchical information of rows
(or columns or both) of \code{assays} tables.
}
\details{
The class \strong{TreeSummarizedExperiment} is designed to store
  rectangular data for entities (e.g., microbes or cell types)
  (\code{assays}), information about the hiearchical structure
  (\code{treeData}), and the mapping information between the rows (or
  columns, or both) of the rectangular data and the tree nodes
  (\code{linkData}). Users could provide hiearchical information on the rows
  or columns (or both) of the \code{assay} tables, and the \code{linkData}
  will be automatically integrated as one part of the \code{rowData} or
  \code{colData} or both, respectively. When the \code{linkData} is added to
  \code{rowData} or \code{colData}, a class \code{LinkDataFrame} is used to
  store data instead of \code{DataFrame}. Please see the page
  \code{\link{LinkDataFrame}} for more details.
}
\section{Slots}{

\describe{
\item{\code{treeData}}{A list of phylo objects. It gives information about the
hiearchical structure of rows or columns of \code{assay} tables.}

\item{\code{...}}{Other slots from
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}}
}}

\section{Constructor}{

See \code{\link{TreeSummarizedExperiment-constructor}} for constructor
functions.
}

\section{Accessor}{

See \code{\link{TreeSummarizedExperiment-accessor}} for accessor functions.
}

\seealso{
\code{\link{TreeSummarizedExperiment}}
  \code{\link{TreeSummarizedExperiment-accessor}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
