% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{TreeSummarizedExperiment-constructor}
\alias{TreeSummarizedExperiment-constructor}
\alias{TreeSummarizedExperiment}
\title{Construct a TreeSummarizedExperiment object}
\usage{
TreeSummarizedExperiment(rowTree = NULL, colTree = NULL, ...)
}
\arguments{
\item{rowTree}{A phylo object that provides hiearchical information of rows
of assay tables.}

\item{colTree}{A phylo object that provides hiearchical information of
columns of assay tables.}

\item{...}{For \code{SummarizedExperiment}, S4 methods \code{list}
    and \code{matrix}, arguments identical to those of the
    \code{SimpleList} method.

    For \code{rowRanges}, ignored.

  }
}
\value{
a TreeSummarizedExperiment object
}
\description{
\code{TreeSummarizedExperiment} constructs a TreeSummarizedExperiment object.
}
\examples{
data("tinyTree")

# the count table
count <- matrix(rpois(100, 50), nrow = 10)
rownames(count) <- c(tinyTree$tip.label)
colnames(count) <- paste("C_", 1:10, sep = "_")

# The sample information
sampC <- data.frame(condition = rep(c("control", "trt"), each = 5),
                    gender = sample(x = 1:2, size = 10, replace = TRUE))
rownames(sampC) <- colnames(count)

# build a TreeSummarizedExperiment object
tse <- TreeSummarizedExperiment(rowTree = tinyTree,
                                assays = list(count),
                                colData = sampC)

}
\seealso{
\code{\link{TreeSummarizedExperiment-class}}
  \code{\link{TreeSummarizedExperiment-accessor}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
