% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperiment.R
\name{TreeSummarizedExperiment-internal}
\alias{TreeSummarizedExperiment-internal}
\title{TreeSummarizedExperiment internals}
\description{
Internal functions which should only be used in TreeSummarizedExperiment
}
\keyword{internal}
