% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject.R
\name{updateObject,TreeSummarizedExperiment-method}
\alias{updateObject,TreeSummarizedExperiment-method}
\title{Update a \code{TreeSummarizedExperiment} object}
\usage{
\S4method{updateObject}{TreeSummarizedExperiment}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A \code{TreeSummarizedExperiment} object}

\item{...}{additional arguments, for use in specific \code{updateObject}
methods.}

\item{verbose}{\code{TRUE} or \code{FALSE}, indicating whether information 
about the update should be reported.}
}
\value{
An updated \code{TreeSummarizedExperiment} object
}
\description{
Update \code{TreeSummarizedExperiment} objects to the latest version of the 
class structure. This is usually called by methods in the 
\code{TreeSummarizedExperiment} package rather than by users or downstream 
packages.
}
