% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggValue.R
\docType{methods}
\name{aggValue}
\alias{aggValue}
\alias{aggValue,matrix-method}
\alias{aggValue,TreeSummarizedExperiment-method}
\title{Perform data aggregations based on the available tree structures}
\usage{
aggValue(x, rowTree = NULL, colTree = NULL, rowLevel = NULL,
  colLevel = NULL, FUN = sum, message = FALSE)

\S4method{aggValue}{matrix}(x, rowTree = NULL, colTree = NULL,
  rowLevel = NULL, colLevel = NULL, FUN = sum, message = FALSE)

\S4method{aggValue}{TreeSummarizedExperiment}(x, rowTree = NULL,
  colTree = NULL, rowLevel = NULL, colLevel = NULL, FUN = sum,
  message = FALSE)
}
\arguments{
\item{x}{A \code{TreeSummarizedExperiment} object or a matrix. If the latter
is given, the tree structure is required (more details in \code{rowTree}
and \code{colTree}).}

\item{rowTree}{An optional argument. The tree structure on the rows of the
matrix. If required, a \code{phylo} object should be provided.}

\item{colTree}{An optional argument. The tree structure on the columns of the
matrix. If required, a \code{phylo} object should be provided.}

\item{rowLevel}{A numeric or character vector. The default is NULL. It
provides the level on the tree that the aggregation is performed to. The
aggregation is on the row dimension of matrix (or the \code{assay} tables).}

\item{colLevel}{A numeric or character vector. The default is NULL. It
provides the level on the tree that the aggregation is performed to. The
aggregation is on the column dimension of matrix (or the \code{assay}
tables).}

\item{FUN}{A function to be applied on the aggregation. It's similar to the
\code{FUN} in \code{\link[base]{apply}}}

\item{message}{A logical value. The default is TRUE. If TRUE, it will print
out the running process.}
}
\value{
A \code{TreeSummarizedExperiment} object or a
  \code{matrix}. The output has the same class of the input \code{x}.
}
\description{
\code{aggValue} aggregates values on the leaf nodes of a tree to a specific
arbitrary level of the tree. The level is specified via the nodes of the
tree. Users could decide on which dimension (row or column) and how should
the aggregation be performed.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Aggregation with input x as a matrix

\item \code{TreeSummarizedExperiment}: Aggregation with input x as a TreeSummarizedExperiment
}}

\examples{

# assays data
set.seed(1)
toyTable <- matrix(rnbinom(20, size = 1, mu = 10), nrow = 5)
colnames(toyTable) <- paste(rep(LETTERS[1:2], each = 2),
                            rep(1:2, 2), sep = "_")
rownames(toyTable) <- paste("entity", seq_len(5), sep = "")

toyTable

# the column data
colInf <- DataFrame(gg = c(1, 2, 3, 3),
                    group = rep(LETTERS[1:2], each = 2),
                    row.names = colnames(toyTable))
colInf

# the toy tree
library(ape)
set.seed(4)
treeC <- rtree(4)
treeC$node.label <- c("All", "GroupA", "GroupB")

library(ggtree)
ggtree(treeC, size = 2) +
   geom_text2(aes(label = node), color = "darkblue",
           hjust = -0.5, vjust = 0.7, size = 6) +
    geom_text2(aes(label = label), color = "darkorange",
               hjust = -0.1, vjust = -0.7, size = 6)

colInf$nodeLab <- treeC$tip.label
tse <- TreeSummarizedExperiment(assays = list(toyTable),
                                colData = colInf,
                                colTree = treeC)

aggRow <- aggValue(x = tse, colLevel = c("GroupA", "GroupB"),
FUN = sum)

assays(aggRow)[[1]]
}
\seealso{
\code{\link{TreeSummarizedExperiment}}
}
\author{
Ruizhu HUANG
}
