% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{LinkDataFrame-class}
\alias{LinkDataFrame-class}
\title{LinkDataFrame: A S4 class extended from DataFrame
An S4 class LinkDataFrame}
\description{
The \strong{LinkDataFrame} is extended from the class \strong{DataFrame} to
include at least four columns \code{nodeLab}, \code{nodeLab_alias},
\code{nodeNum}, and \code{isLeaf}.
}
\section{Constructor}{

See \code{\link{LinkDataFrame-constructor}} for constructor
functions.
}

