% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{LinkDataFrame}
\alias{LinkDataFrame}
\title{Construct a LinkDataFrame
Construct a LinkDataFrame object}
\usage{
LinkDataFrame(LinkData = NULL, ...)
}
\arguments{
\item{LinkData}{A \link[S4Vectors]{DataFrame-class}.}

\item{...}{For \code{SummarizedExperiment}, S4 methods \code{list}
    and \code{matrix}, arguments identical to those of the
    \code{SimpleList} method.

    For \code{rowRanges}, ignored.

  }
}
\value{
A LinkDataFrame object
}
\description{
Construct a LinkDataFrame
Construct a LinkDataFrame object
}
\examples{

left <- DataFrame(left1 = 1:5, left2 = letters[1:5])
right <- DataFrame(right1 = sample(letters[1:3], 5, replace = TRUE),
                  right2 = sample(c(TRUE, FALSE), 5, replace = TRUE),
                  right3 = 11:15)

(ld <- LinkDataFrame(LinkData = left, right))
}
\seealso{
\code{\link{LinkDataFrame-accessor}}
  \code{\link{LinkDataFrame-class}}
}
