% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{LinkDataFrame-constructor}
\alias{LinkDataFrame-constructor}
\alias{LinkDataFrame}
\title{Construct a LinkDataFrame
Construct a LinkDataFrame object}
\usage{
LinkDataFrame(nodeLab, nodeLab_alias, nodeNum, isLeaf, ...)
}
\arguments{
\item{nodeLab}{A character vector}

\item{nodeLab_alias}{A character vector}

\item{nodeNum}{A numeric vector}

\item{isLeaf}{A logical vector}

\item{...}{All arguments accepted by \code{\link[S4Vectors]{DataFrame-class}}.}
}
\value{
A LinkDataFrame object
}
\description{
Construct a LinkDataFrame
Construct a LinkDataFrame object
}
\examples{


(ld <- LinkDataFrame(nodeLab = letters[1:5],
                     nodeLab_alias = LETTERS[1:5],
                     nodeNum = 1:5,
                     isLeaf = TRUE,
                     right = 1:5))
}
\seealso{
\code{\link{LinkDataFrame-class}}
  \code{\link[S4Vectors]{DataFrame-class}}
}
