% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTSE.R
\name{makeTSE}
\alias{makeTSE}
\title{A toy TreeSummarizedExperiment}
\usage{
makeTSE(nrow = 10, ncol = 4, include.rowTree = TRUE, include.colTree = TRUE)
}
\arguments{
\item{nrow}{a numeric value to specify the number of rows of
\code{TreeSummarizedExperiment}}

\item{ncol}{a numeric value to specify the number of columns of
\code{TreeSummarizedExperiment}}

\item{include.rowTree}{TRUE or FALSE. Default is TRUE, so the output
\code{TreeSummarizedExperiment} has a \code{phylo} object in
\code{rowTree}.}

\item{include.colTree}{TRUE or FALSE. Default is TRUE, so the output
\code{TreeSummarizedExperiment} has a \code{phylo} object in
\code{colTree}.}
}
\value{
A TreeSummarizedExperiment object
}
\description{
\code{makeTSE} creates a toy TreeSummarizedExperiment
}
\examples{

set.seed(1)
makeTSE()
}
\author{
Ruizhu Huang
}
