% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTree.R
\name{toTree}
\alias{toTree}
\title{Translate a data frame to a phylo object}
\usage{
toTree(data)
}
\arguments{
\item{data}{A data frame or matrix.}
}
\value{
a phylo object
}
\description{
\code{toTree} translates a data frame to a phylo object
}
\examples{

taxTab <- data.frame(R1 = rep("A", 5),
R2 = c("B1", rep("B2", 4)),
R3 = c("C1", "C2", "C3", "C3", "C4"))


tree <- toTree(data = taxTab)

}
\author{
Ruizhu HUANG
}
