% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/classAccessor.R
\docType{methods}
\name{linkData}
\alias{linkData}
\alias{linkData,TreeSummarizedExperiment-method}
\alias{treeData}
\alias{treeData,TreeSummarizedExperiment-method}
\alias{TreeSummarizedExperiment-accessor}
\title{TreeSummarizedExperiment-accessors}
\usage{
linkData(x, onRow = TRUE)

\S4method{linkData}{TreeSummarizedExperiment}(x, onRow = TRUE)

treeData(x, onRow = TRUE)

\S4method{treeData}{TreeSummarizedExperiment}(x, onRow = TRUE)
}
\arguments{
\item{x}{A TreeSummarizedExperiment object}

\item{onRow}{A logical(1) value, TRUE or FALSE. The default is TRUE. It
indicates the returned \code{treeData} or \code{linkData} is on rows (TRUE)
or columns (FALSE) of \code{assays} tables.}
}
\value{
Elements from \code{TreeSummarizedExperiment}.
}
\description{
All accessor functions that work on
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} should work on
\strong{TreeSummarizedExperiment}. Additionally, two new \code{linkData} and
\code{treeData} accessor function are available for
\strong{TreeSummarizedExperiment}.
}
\examples{

# the assay table
set.seed(1)
y <- matrix(rnbinom(300,size=1,mu=10),nrow=10)
colnames(y) <- paste(rep(LETTERS[1:3], each = 10), rep(1:10,3), sep = "_")
rownames(y) <- tinyTree$tip.label

# the row data
rowInf <- DataFrame(var1 = sample(letters[1:3], 10, replace = TRUE),
                    var2 = sample(c(TRUE, FALSE), 10, replace = TRUE))
# the column data
colInf <- DataFrame(gg = factor(sample(1:3, 30, replace = TRUE)),
                    group = rep(LETTERS[1:3], each = 10))

# the tree structure on the rows of assay tables
data("tinyTree")

# the tree structure on the columns of assay tables
sampTree <- ape::rtree(30)
sampTree$tip.label <- colnames(y)

# create the TreeSummarizedExperiment object
rowInf$nodeLab <- rownames(y)
colInf$nodeLab <- colnames(y)
toy_tse <- TreeSummarizedExperiment(assays = list(y),
                                    rowData = rowInf,
                                    colData = colInf,
                                    rowTree = tinyTree,
                                    colTree = sampTree)

## extract the rowData
(rowD <- rowData(x = toy_tse))

## extract the colData
(colD <- colData(x = toy_tse))

## extract the linkData
# on rows
(rowL <- linkData(x = toy_tse, onRow = TRUE))
# on columns
(colL <- linkData(x = toy_tse, onRow = FALSE))

 ## extract the treeData
# on rows
(rowT <- treeData(x = toy_tse, onRow = TRUE))
# on columns
(colT <- treeData(x = toy_tse, onRow = FALSE))

}
\seealso{
\code{\link{TreeSummarizedExperiment}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
