% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/classAccessor.R
\name{rowLinks}
\alias{rowLinks}
\alias{rowLinks,TreeSummarizedExperiment-method}
\alias{colLinks}
\alias{colLinks,TreeSummarizedExperiment-method}
\alias{rowTree}
\alias{rowTree,TreeSummarizedExperiment-method}
\alias{colTree}
\alias{colTree,TreeSummarizedExperiment-method}
\alias{[,TreeSummarizedExperiment,ANY,ANY,ANY-method}
\alias{rownames<-,TreeSummarizedExperiment-method}
\alias{colnames<-,TreeSummarizedExperiment-method}
\alias{TreeSummarizedExperiment-accessor}
\title{TreeSummarizedExperiment-accessors}
\usage{
rowLinks(x)

\S4method{rowLinks}{TreeSummarizedExperiment}(x)

colLinks(x)

\S4method{colLinks}{TreeSummarizedExperiment}(x)

rowTree(x)

\S4method{rowTree}{TreeSummarizedExperiment}(x)

colTree(x)

\S4method{colTree}{TreeSummarizedExperiment}(x)

\S4method{[}{TreeSummarizedExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{rownames}{TreeSummarizedExperiment}(x) <- value

\S4method{colnames}{TreeSummarizedExperiment}(x) <- value
}
\arguments{
\item{x}{A TreeSummarizedExperiment object}

\item{i, j}{The row, column index to subset \code{x}. The arguments of the
subset function \code{[]}}

\item{...}{The argument from the subset function \code{[]}}

\item{drop}{A logical value, TRUE or FALSE. The argument from the subset
function \code{[]}}

\item{value}{the new rownames or colnames as a \code{character} value. 
See \code{\link[BiocGenerics:row_colnames]{BiocGenerics}}.}
}
\value{
Elements from \code{TreeSummarizedExperiment}.
}
\description{
All accessor functions that work on
\code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}} should work on
\strong{TreeSummarizedExperiment}. Additionally, new accessors
\code{rowLinks} \code{colLinks}, \code{rowTree} and \code{colTree} accessor
function are available for \strong{TreeSummarizedExperiment}.
}
\examples{

# the assay table
set.seed(1)
y <- matrix(rnbinom(300,size=1,mu=10),nrow=10)
colnames(y) <- paste(rep(LETTERS[1:3], each = 10), rep(1:10,3), sep = "_")
rownames(y) <- tinyTree$tip.label

# the row data
rowInf <- DataFrame(var1 = sample(letters[1:3], 10, replace = TRUE),
                    var2 = sample(c(TRUE, FALSE), 10, replace = TRUE))
# the column data
colInf <- DataFrame(gg = factor(sample(1:3, 30, replace = TRUE)),
                    group = rep(LETTERS[1:3], each = 10))

# the tree structure on the rows of assay tables
data("tinyTree")

# the tree structure on the columns of assay tables
sampTree <- ape::rtree(30)
sampTree$tip.label <- colnames(y)

# create the TreeSummarizedExperiment object
toy_tse <- TreeSummarizedExperiment(assays = list(y),
                                    rowData = rowInf,
                                    colData = colInf,
                                    rowTree = tinyTree,
                                    colTree = sampTree)

## extract the rowData
(rowD <- rowData(x = toy_tse))

## extract the colData
(colD <- colData(x = toy_tse))

## extract the linkData
# on rows
(rowL <- rowLinks(x = toy_tse))
# on columns
(colL <- colLinks(x = toy_tse))

 ## extract the treeData
# on rows
(rowT <- rowTree(x = toy_tse))
# on columns
(colT <- colTree(x = toy_tse))

}
\seealso{
\code{\link{TreeSummarizedExperiment}}
  \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}}
}
\author{
Ruizhu HUANG
}
