% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{TreeSummarizedExperiment-class}
\alias{TreeSummarizedExperiment-class}
\title{An S4 class TreeSummarizedExperiment}
\description{
The class \strong{TreeSummarizedExperiment} is an extension class of standard
\code{\link[SingleCellExperiment]{SingleCellExperiment-class}} class. It has
four more slots that are not in
\code{\link[SingleCellExperiment]{SingleCellExperiment-class}} class:
\code{rowTree}, \code{rowLink} \code{colTree} and \code{colLink}. The
hierarchical information of rows (columns) is stored in \code{rowTree}
(\code{colTree}) and the link between the rows (columns) of \code{assays}
tables and nodes of the tree is given in \code{rowLink} (\code{colLink}).
}
\details{
The class \strong{TreeSummarizedExperiment} is designed to store
  rectangular data for entities (e.g., microbes or cell types)
  (\code{assays}), information about the hiearchical structure
  (\code{rowTree} on rows; \code{colTree} on columns), and the mapping
  information between the tree nodes and the rows or the columns of the
  rectangular data. Users could provide the hiearchical structure of the
  rows, columns or both) of the \code{assay} tables, and the link data will
  be automatically generated in \code{rowLink}, \code{colData} or both,
  respectively. It's required that the object in \code{rowLink} or
  \code{colLink} has the \code{LinkDataFrame} class. Please see the page
  \code{\link{LinkDataFrame}} for more details.
}
\section{Slots}{

\describe{
\item{\code{rowTree}}{A phylo object or NULL. It gives information about the
hiearchical structure of rows of \code{assay} tables.}

\item{\code{colTree}}{A phylo object or NULL. It gives information about the
hiearchical structure of columns of \code{assay} tables.}

\item{\code{rowLink}}{A LinkDataFrame. It gives information about the link between
the nodes of the \code{rowTree} and the rows of \code{assay} tables.}

\item{\code{colLink}}{A LinkDataFrame. It gives information about the link between
the nodes of the \code{colTree} and the columns of \code{assay} tables.}

\item{\code{...}}{Other slots from
\code{\link[SingleCellExperiment]{SingleCellExperiment-class}}}
}}

\section{Constructor}{

See \code{\link{TreeSummarizedExperiment-constructor}} for constructor
functions.
}

\section{Accessor}{

See \code{\link{TreeSummarizedExperiment-accessor}} for accessor functions.
}

\seealso{
\code{\link{TreeSummarizedExperiment}}
  \code{\link{TreeSummarizedExperiment-accessor}}
  \code{\link[SingleCellExperiment]{SingleCellExperiment-class}}
}
