% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{treeSummarizedExperiment-constructor}
\alias{treeSummarizedExperiment-constructor}
\alias{treeSummarizedExperiment}
\title{Construct a treeSummarizedExperiment object}
\usage{
treeSummarizedExperiment(tree = NULL, linkData = NULL, ...)
}
\arguments{
\item{tree}{A phylo object}

\item{linkData}{A data frame or \code{\link[S4Vectors]{DataFrame-class}}. It
has the same number of rows as the matrix-like elements of \code{assays}.
The row order of the \code{linkData} matches with that of  the matrix-like
element of \code{assays}. It has the following columns.
\itemize{
\item \strong{nodeLab} The labels of nodes on the tree.
\item \strong{nodeLab_alias} An alias of column \code{nodeLab}. It is
created only when there are missing value or duplicated value in column
\code{nodeLab}. A prefix "Node_" and "Leaf_" is added to the node number
(column \code{nodeNum}) for the internal nodes and the leaf nodes,
respectively.
\item \strong{nodeNum} The numbers of nodes
\item \strong{isLeaf} It indicates whether the node is a leaf node
\item \strong{rowID} The corresponding row number of the matrix-like
elements in \code{assays} }}

\item{...}{For \code{SummarizedExperiment}, S4 methods \code{list}
    and \code{matrix}, arguments identical to those of the
    \code{SimpleList} method.

    For \code{rowRanges}, ignored.

  }
}
\value{
a treeSummarizedExperiment object
}
\description{
\code{treeSummarizedExperiment} constructs a treeSummarizedExperiment object.
}
\examples{
data("tinyTree")

# the count table
count <- matrix(rpois(100, 50), nrow = 10)
rownames(count) <- c(tinyTree$tip.label)
colnames(count) <- paste("C_", 1:10, sep = "_")

# The sample information
sampC <- data.frame(condition = rep(c("control", "trt"), each = 5),
                    gender = sample(x = 1:2, size = 10, replace = TRUE))
rownames(sampC) <- colnames(count)

# build a treeSummarizedExperiment object
tse <- treeSummarizedExperiment(tree = tinyTree,
                                assays = list(count),
                                colData = sampC)

}
\seealso{
\code{\link{treeSummarizedExperiment-class}}
  \code{\link{treeSummarizedExperiment-accessor}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
