% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSummarizedExperiment.R, R/allClass.R
\name{TreeSummarizedExperiment-internal}
\alias{TreeSummarizedExperiment-internal}
\alias{vertical_slot_names,TreeSummarizedExperiment-method}
\title{TreeSummarizedExperiment internals}
\usage{
\S4method{vertical_slot_names}{TreeSummarizedExperiment}(x)
}
\description{
Internal functions which should only be used in TreeSummarizedExperiment
}
\keyword{internal}
